# contracts/v1/management.openapi.yaml
openapi: 3.1.0
info:
  title: Liapoldus Gateway Management API
  version: 1.0.0
  description: >-
    Control plane for plugin instances, desired configuration, access,
    operations and audit. Product-specific runtime administration is exposed
    only by the corresponding plugin Admin Surface.
servers:
  - url: https://gateway-management.example.internal
security:
  - bearerAuth: []
x-liapoldus-security:
  remoteAccess: private-network-or-vpn-only
  webConstructorBackendRequired: [mtls, bearerAuth]
  desktopAccess: ssh-port-forward-to-loopback
  desktopTunnelRequired: [tlsServerVerification, bearerAuth]
  loopbackRequired: [bearerAuth]
  roleModel: platform-admin-only
  humanUserAuthorization: Constructor-backend-only
  perGatewayCredential: true
  managementAndPluginTrustRootsMustDiffer: true
paths:
  /healthz:
    get:
      operationId: health
      security: []
      responses:
        '200': { description: Process health, content: { application/json: { schema: { $ref: '#/components/schemas/Health' } } } }
  /api/status:
    get:
      operationId: getStatus
      responses:
        '200': { description: Gateway control-plane readiness, content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } } }
  /api/plugins:
    get:
      operationId: listPlugins
      responses:
        '200': { description: Configured plugin instances, content: { application/json: { schema: { $ref: '#/components/schemas/PluginList' } } } }
    post:
      operationId: createPlugin
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PluginInput' } } } }
      responses:
        '201': { description: Created plugin instance, content: { application/json: { schema: { $ref: '#/components/schemas/Plugin' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /api/plugins/{pluginId}:
    get:
      operationId: getPlugin
      parameters: [{ $ref: '#/components/parameters/PluginId' }]
      responses:
        '200': { description: Plugin resource, content: { application/json: { schema: { $ref: '#/components/schemas/Plugin' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      operationId: updatePlugin
      parameters: [{ $ref: '#/components/parameters/PluginId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PluginInput' } } } }
      responses:
        '200': { description: Updated plugin, content: { application/json: { schema: { $ref: '#/components/schemas/Plugin' } } } }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deletePlugin
      parameters: [{ $ref: '#/components/parameters/PluginId' }]
      responses:
        '204': { description: Plugin instance removed. }
        '409': { $ref: '#/components/responses/Problem' }
  /api/plugins/{pluginId}/restart:
    post:
      operationId: restartPlugin
      parameters: [{ $ref: '#/components/parameters/PluginId' }]
      responses:
        '202': { description: Restart local process or refresh remote connection; remote process is not restarted by Gateway., content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } }
  /api/operations/{operationId}:
    get:
      operationId: getOperation
      parameters: [{ $ref: '#/components/parameters/OperationId' }]
      responses:
        '200': { description: Durable operation state, content: { application/json: { schema: { $ref: '#/components/schemas/Operation' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /api/audit:
    get:
      operationId: listAudit
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: Redacted audit events, content: { application/json: { schema: { $ref: '#/components/schemas/AuditList' } } } }
  /api/access/service-keys:
    get:
      operationId: listServiceKeys
      responses:
        '200': { description: Key metadata only; verifier and credential material are excluded., content: { application/json: { schema: { $ref: '#/components/schemas/ServiceKeyList' } } } }
    post:
      operationId: createServiceKey
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceKeyInput' } } } }
      responses:
        '201': { description: Raw credential returned once, content: { application/json: { schema: { $ref: '#/components/schemas/CreatedServiceKey' } } } }
  /api/access/service-keys/{keyId}/rotate:
    post:
      operationId: rotateServiceKey
      parameters: [{ $ref: '#/components/parameters/KeyId' }]
      responses:
        '200': { description: New raw credential returned once., content: { application/json: { schema: { $ref: '#/components/schemas/CreatedServiceKey' } } } }
  /api/access/service-keys/{keyId}/revoke:
    post:
      operationId: revokeServiceKey
      parameters: [{ $ref: '#/components/parameters/KeyId' }]
      responses:
        '204': { description: Revoked key metadata; no credential is returned. }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: Gateway platform-admin service token }
    mtls: { type: mutualTLS, description: Required with Bearer on every remote Management API listener. }
  parameters:
    PluginId: { name: pluginId, in: path, required: true, schema: { type: string, pattern: '^[a-z][a-z0-9-]{0,62}$' } }
    OperationId: { name: operationId, in: path, required: true, schema: { type: string } }
    KeyId: { name: keyId, in: path, required: true, schema: { type: string } }
    Cursor: { name: cursor, in: query, schema: { type: string, description: Opaque cursor } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string } }
  responses:
    Problem: { description: RFC 9457 problem, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  schemas:
    Health: { type: object, required: [status, requestId], properties: { status: { const: ok }, requestId: { type: string } } }
    Status: { type: object, required: [drift, dataPlaneReadiness, requestId], properties: { drift: { type: boolean }, dataPlaneReadiness: { $ref: '#/components/schemas/DataPlaneReadiness' }, requestId: { type: string } } }
    DataPlaneReadiness: { type: object, required: [state], properties: { state: { enum: [ready, not-ready] }, reason: { enum: [system-release-required, plugin-unavailable, recovery-required], description: Present only when state is not-ready. } } }
    PluginInput: { type: object, additionalProperties: false, required: [mode, settings], properties: { mode: { enum: [local, remote] }, binary: { type: string }, endpoint: { type: string }, tls: { type: object }, settings: { type: object }, grants: { type: object }, limits: { type: object } } }
    Plugin: { type: object, required: [id, mode, state, capabilities, settings], properties: { id: { type: string }, mode: { enum: [local, remote] }, state: { enum: [starting, ready, unhealthy, stopped] }, endpoint: { type: string }, capabilities: { type: array, items: { type: string } }, capabilityModes: { type: object, description: Invocation modes attested by the plugin Manifest., additionalProperties: { type: array, items: { enum: [call, http-stream, websocket, sse, tcp, udp] } } }, settings: { type: object, description: Redacted settings. } } }
    PluginList: { type: object, required: [items, requestId], properties: { items: { type: array, items: { $ref: '#/components/schemas/Plugin' } }, requestId: { type: string } } }
    IdempotencyRequest: { type: object, additionalProperties: false, required: [idempotencyKey], properties: { idempotencyKey: { type: string, minLength: 16, maxLength: 128 } } }
    OperationReference: { type: object, required: [operationId, state, requestId], properties: { operationId: { type: string }, state: { enum: [pending, running] }, requestId: { type: string } } }
    Operation: { type: object, required: [id, kind, state, createdAt, requestId], properties: { id: { type: string }, kind: { type: string }, state: { enum: [pending, running, succeeded, failed] }, createdAt: { type: string, format: date-time }, updatedAt: { type: string, format: date-time }, result: {}, problem: { $ref: '#/components/schemas/Problem' }, requestId: { type: string } } }
    AuditList: { type: object, required: [items, nextCursor, requestId], properties: { items: { type: array, items: { type: object, required: [timestamp, actor, action, result], properties: { timestamp: { type: string, format: date-time }, actor: { type: string }, action: { type: string }, resource: { type: string }, result: { enum: [succeeded, failed] } } } }, nextCursor: { type: [string, 'null'] }, requestId: { type: string } } }
    ServiceKeyInput: { type: object, additionalProperties: false, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 80 } } }
    CreatedServiceKey: { type: object, required: [id, name, role, token, requestId], properties: { id: { type: string }, name: { type: string }, role: { const: platform-admin }, token: { type: string, writeOnly: true }, requestId: { type: string } } }
    ServiceKeyList: { type: object, additionalProperties: false, required: [items, requestId], properties: { items: { type: array, items: { $ref: '#/components/schemas/ServiceKeyMetadata' } }, requestId: { type: string } } }
    ServiceKeyMetadata: { type: object, additionalProperties: false, required: [id, name, role, createdAt, expiresAt, revokedAt], properties: { id: { type: string }, name: { type: string }, role: { const: platform-admin }, createdAt: { type: string }, expiresAt: { type: [string, 'null'] }, revokedAt: { type: [string, 'null'] } } }
    Problem: { type: object, required: [type, title, status, code, detail, instance, requestId], properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, code: { type: string }, detail: { type: string }, instance: { type: string }, requestId: { type: string }, diagnostics: { type: array, items: { type: object } } } }